Added support of output dataclass
