+++ QEfficient/diffusers/pipelines/pipeline_module.py
@@ -5,7 +5,8 @@
 #
 # ----------------------------------------------------------------------------
 
-from typing import Dict, List, Tuple
+import copy
+import hashlib
 
 import torch
 import torch.nn as nn
@@ -16,107 +17,54 @@ from QEfficient.diffusers.models.pytorch_transforms import (
     AttentionTransform,
     CustomOpsTransform,
     NormalizationTransform,
-)
-from QEfficient.diffusers.models.transformers.transformer_flux import (
-    QEffFluxSingleTransformerBlock,
-    QEffFluxTransformerBlock,
+    OnnxFunctionTransform,
 )
 from QEfficient.transformers.models.pytorch_transforms import (
     T5ModelTransform,
 )
 from QEfficient.utils import constants
+from QEfficient.utils.cache import to_hashable
 
 
 class QEffTextEncoder(QEFFBaseModel):
-    """
-    Wrapper for text encoder models with ONNX export and QAIC compilation capabilities.
-
-    This class handles text encoder models (CLIP, T5) with specific transformations and
-    optimizations for efficient inference on Qualcomm AI hardware. It applies custom
-    PyTorch and ONNX transformations to prepare models for deployment.
-
-    Attributes:
-        model (nn.Module): The wrapped text encoder model (deep copy of original)
-        _pytorch_transforms (List): PyTorch transformations applied before ONNX export
-        _onnx_transforms (List): ONNX transformations applied after export
-    """
-
     _pytorch_transforms = [CustomOpsTransform, T5ModelTransform]
     _onnx_transforms = [FP16ClipTransform, SplitTensorsTransform]
+    """
+    QEffTextEncoder is a wrapper class for text encoder models that provides ONNX export and compilation capabilities.
 
-    @property
-    def get_model_config(self) -> Dict:
-        """
-        Get the model configuration as a dictionary.
+    This class extends QEFFBaseModel to handle text encoder models (like T5EncoderModel) with specific
+    transformations and optimizations for efficient inference on Qualcomm AI hardware.
+    """
 
-        Returns:
-            Dict: The configuration dictionary of the underlying text encoder model
-        """
-        return self.model.config.__dict__
-
-    def __init__(self, model: nn.Module) -> None:
-        """
-        Initialize the text encoder wrapper.
-
-        Args:
-            model (nn.Module): The text encoder model to wrap (CLIP or T5)
-        """
+    def __init__(self, model: nn.modules):
         super().__init__(model)
-        self.model = model
+        self.model = copy.deepcopy(model)
 
-    def get_onnx_params(self) -> Tuple[Dict, Dict, List[str]]:
-        """
-        Generate ONNX export configuration for the text encoder.
-
-        Creates example inputs, dynamic axes specifications, and output names
-        tailored to the specific text encoder type (CLIP vs T5).
-
-        Returns:
-            Tuple containing:
-                - example_inputs (Dict): Sample inputs for ONNX export
-                - dynamic_axes (Dict): Specification of dynamic dimensions
-                - output_names (List[str]): Names of model outputs
-        """
+    def get_onnx_config(self):
         bs = constants.ONNX_EXPORT_EXAMPLE_BATCH_SIZE
 
-        # Create example input with max sequence length
         example_inputs = {
             "input_ids": torch.zeros((bs, self.model.config.max_position_embeddings), dtype=torch.int64),
         }
 
-        # Define which dimensions can vary at runtime
         dynamic_axes = {"input_ids": {0: "batch_size", 1: "seq_len"}}
+        output_names = ["last_hidden_state", "pooler_output"]
 
-        # T5 only outputs hidden states, CLIP outputs both hidden states and pooled output
         if self.model.__class__.__name__ == "T5EncoderModel":
             output_names = ["last_hidden_state"]
         else:
-            output_names = ["last_hidden_state", "pooler_output"]
             example_inputs["output_hidden_states"] = False
 
         return example_inputs, dynamic_axes, output_names
 
     def export(
         self,
-        inputs: Dict,
-        output_names: List[str],
-        dynamic_axes: Dict,
-        export_dir: str = None,
-        export_kwargs: Dict = None,
-    ) -> str:
-        """
-        Export the text encoder model to ONNX format.
-
-        Args:
-            inputs (Dict): Example inputs for ONNX export
-            output_names (List[str]): Names of model outputs
-            dynamic_axes (Dict): Specification of dynamic dimensions
-            export_dir (str, optional): Directory to save ONNX model
-            export_kwargs (Dict, optional): Additional export arguments
-
-        Returns:
-            str: Path to the exported ONNX model
-        """
+        inputs,
+        output_names,
+        dynamic_axes,
+        export_dir=None,
+        export_kwargs=None,
+    ):
         return self._export(
             example_inputs=inputs,
             output_names=output_names,
@@ -125,75 +73,50 @@ class QEffTextEncoder(QEFFBaseModel):
             export_kwargs=export_kwargs,
         )
 
-    def compile(self, specializations: List[Dict], **compiler_options) -> None:
-        """
-        Compile the ONNX model for Qualcomm AI hardware.
-
-        Args:
-            specializations (List[Dict]): Model specialization configurations
-            **compiler_options: Additional compiler options (e.g., num_cores, aic_num_of_activations)
-        """
+    def compile(self, specializations, **compiler_options):
         self._compile(specializations=specializations, **compiler_options)
+
+    @property
+    def model_hash(self) -> str:
+        # Compute the hash with: model_config, continuous_batching, transforms
+        mhash = hashlib.sha256()
+        mhash.update(to_hashable(self.model.config.to_diff_dict()))
+        mhash.update(to_hashable(self._transform_names()))
+        mhash = mhash.hexdigest()[:16]
+        return mhash
+
+    @property
+    def model_name(self) -> str:
+        mname = self.model.__class__.__name__
+        if mname.startswith("QEff") or mname.startswith("QEFF"):
+            mname = mname[4:]
+        return mname
 
 
 class QEffUNet(QEFFBaseModel):
-    """
-    Wrapper for UNet models with ONNX export and QAIC compilation capabilities.
-
-    This class handles UNet models with specific transformations and optimizations
-    for efficient inference on Qualcomm AI hardware. UNet is commonly used in
-    diffusion models for image generation tasks.
-
-    Attributes:
-        model (nn.Module): The wrapped UNet model
-        _pytorch_transforms (List): PyTorch transformations applied before ONNX export
-        _onnx_transforms (List): ONNX transformations applied after export
-    """
-
     _pytorch_transforms = [CustomOpsTransform]
     _onnx_transforms = [FP16ClipTransform, SplitTensorsTransform]
 
-    @property
-    def get_model_config(self) -> Dict:
-        """
-        Get the model configuration as a dictionary.
+    """
+    QEffUNet is a wrapper class for UNet models that provides ONNX export and compilation capabilities.
 
-        Returns:
-            Dict: The configuration dictionary of the underlying UNet model
-        """
-        return self.model.config.__dict__
+    This class extends QEFFBaseModel to handle UNet models with specific transformations and optimizations
+    for efficient inference on Qualcomm AI hardware. It is commonly used in diffusion models for image
+    generation tasks.
+    """
 
-    def __init__(self, model: nn.Module) -> None:
-        """
-        Initialize the UNet wrapper.
-
-        Args:
-            model (nn.Module): The pipeline model containing the UNet
-        """
+    def __init__(self, model: nn.modules):
         super().__init__(model.unet)
         self.model = model.unet
 
     def export(
         self,
-        inputs: Dict,
-        output_names: List[str],
-        dynamic_axes: Dict,
-        export_dir: str = None,
-        export_kwargs: Dict = None,
-    ) -> str:
-        """
-        Export the UNet model to ONNX format.
-
-        Args:
-            inputs (Dict): Example inputs for ONNX export
-            output_names (List[str]): Names of model outputs
-            dynamic_axes (Dict): Specification of dynamic dimensions
-            export_dir (str, optional): Directory to save ONNX model
-            export_kwargs (Dict, optional): Additional export arguments
-
-        Returns:
-            str: Path to the exported ONNX model
-        """
+        inputs,
+        output_names,
+        dynamic_axes,
+        export_dir=None,
+        export_kwargs=None,
+    ):
         return self._export(
             example_inputs=inputs,
             output_names=output_names,
@@ -202,76 +125,46 @@ class QEffUNet(QEFFBaseModel):
             export_kwargs=export_kwargs,
         )
 
-    def compile(self, specializations: List[Dict], **compiler_options) -> None:
-        """
-        Compile the ONNX model for Qualcomm AI hardware.
-
-        Args:
-            specializations (List[Dict]): Model specialization configurations
-            **compiler_options: Additional compiler options
-        """
+    def compile(self, specializations, **compiler_options):
         self._compile(specializations=specializations, **compiler_options)
+
+    @property
+    def model_hash(self) -> str:
+        # Compute the hash with: model_config, continuous_batching, transforms
+        mhash = hashlib.sha256()
+        mhash.update(to_hashable(dict(self.model.config)))
+        mhash.update(to_hashable(self._transform_names()))
+        mhash = mhash.hexdigest()[:16]
+        return mhash
+
+    @property
+    def model_name(self) -> str:
+        mname = self.model.__class__.__name__
+        if mname.startswith("QEff") or mname.startswith("QEFF"):
+            mname = mname[4:]
+        return mname
 
 
 class QEffVAE(QEFFBaseModel):
-    """
-    Wrapper for Variational Autoencoder (VAE) models with ONNX export and QAIC compilation.
-
-    This class handles VAE models with specific transformations and optimizations
-    for efficient inference on Qualcomm AI hardware. VAE models are used in diffusion
-    pipelines for encoding images to latent space and decoding latents back to images.
-
-    Attributes:
-        model (nn.Module): The wrapped VAE model (deep copy of original)
-        type (str): VAE operation type ("encoder" or "decoder")
-        _pytorch_transforms (List): PyTorch transformations applied before ONNX export
-        _onnx_transforms (List): ONNX transformations applied after export
-    """
-
     _pytorch_transforms = [CustomOpsTransform]
     _onnx_transforms = [FP16ClipTransform, SplitTensorsTransform]
 
-    @property
-    def get_model_config(self) -> Dict:
-        """
-        Get the model configuration as a dictionary.
+    """
+    QEffVAE is a wrapper class for Variational Autoencoder (VAE) models that provides ONNX export and compilation capabilities.
 
-        Returns:
-            Dict: The configuration dictionary of the underlying VAE model
-        """
-        return self.model.config.__dict__
+    This class extends QEFFBaseModel to handle VAE models with specific transformations and optimizations
+    for efficient inference on Qualcomm AI hardware. VAE models are commonly used in diffusion pipelines
+    for encoding images to latent space and decoding latent representations back to images.
+    """
 
-    def __init__(self, model: nn.Module, type: str) -> None:
-        """
-        Initialize the VAE wrapper.
+    def __init__(self, model: nn.modules, type: str):
+        super().__init__(model.vae)
+        self.model = copy.deepcopy(model.vae)
+        self.type = type
 
-        Args:
-            model (nn.Module): The pipeline model containing the VAE
-            type (str): VAE operation type ("encoder" or "decoder")
-        """
-        super().__init__(model)
-        self.model = model
-
-        # To have different hashing for encoder/decoder
-        self.model.config["type"] = type
-
-    def get_onnx_params(self, latent_height: int = 32, latent_width: int = 32) -> Tuple[Dict, Dict, List[str]]:
-        """
-        Generate ONNX export configuration for the VAE decoder.
-
-        Args:
-            latent_height (int): Height of latent representation (default: 32)
-            latent_width (int): Width of latent representation (default: 32)
-
-        Returns:
-            Tuple containing:
-                - example_inputs (Dict): Sample inputs for ONNX export
-                - dynamic_axes (Dict): Specification of dynamic dimensions
-                - output_names (List[str]): Names of model outputs
-        """
+    def get_onnx_config(self, latent_height=32, latent_width=32):
+        # VAE decode
         bs = constants.ONNX_EXPORT_EXAMPLE_BATCH_SIZE
-
-        # VAE decoder takes latent representation as input
         example_inputs = {
             "latent_sample": torch.randn(bs, 16, latent_height, latent_width),
             "return_dict": False,
@@ -279,34 +172,19 @@ class QEffVAE(QEFFBaseModel):
 
         output_names = ["sample"]
 
-        # All dimensions except channels can be dynamic
         dynamic_axes = {
             "latent_sample": {0: "batch_size", 1: "channels", 2: "latent_height", 3: "latent_width"},
         }
-
         return example_inputs, dynamic_axes, output_names
 
     def export(
         self,
-        inputs: Dict,
-        output_names: List[str],
-        dynamic_axes: Dict,
-        export_dir: str = None,
-        export_kwargs: Dict = None,
-    ) -> str:
-        """
-        Export the VAE model to ONNX format.
-
-        Args:
-            inputs (Dict): Example inputs for ONNX export
-            output_names (List[str]): Names of model outputs
-            dynamic_axes (Dict): Specification of dynamic dimensions
-            export_dir (str, optional): Directory to save ONNX model
-            export_kwargs (Dict, optional): Additional export arguments
-
-        Returns:
-            str: Path to the exported ONNX model
-        """
+        inputs,
+        output_names,
+        dynamic_axes,
+        export_dir=None,
+        export_kwargs=None,
+    ):
         return self._export(
             example_inputs=inputs,
             output_names=output_names,
@@ -315,81 +193,99 @@ class QEffVAE(QEFFBaseModel):
             export_kwargs=export_kwargs,
         )
 
-    def compile(self, specializations: List[Dict], **compiler_options) -> None:
-        """
-        Compile the ONNX model for Qualcomm AI hardware.
-
-        Args:
-            specializations (List[Dict]): Model specialization configurations
-            **compiler_options: Additional compiler options
-        """
+    def compile(self, specializations, **compiler_options):
         self._compile(specializations=specializations, **compiler_options)
+
+    @property
+    def model_hash(self) -> str:
+        # Compute the hash with: model_config, continuous_batching, transforms
+        mhash = hashlib.sha256()
+        mhash.update(to_hashable(dict(self.model.config)))
+        mhash.update(to_hashable(self._transform_names()))
+        mhash.update(to_hashable(self.type))
+        mhash = mhash.hexdigest()[:16]
+        return mhash
+
+    @property
+    def model_name(self) -> str:
+        mname = self.model.__class__.__name__
+        if mname.startswith("QEff") or mname.startswith("QEFF"):
+            mname = mname[4:]
+        return mname
+
+
+class QEffSafetyChecker(QEFFBaseModel):
+    _pytorch_transforms = [CustomOpsTransform]
+    _onnx_transforms = [FP16ClipTransform, SplitTensorsTransform]
+
+    """
+    QEffSafetyChecker is a wrapper class for safety checker models that provides ONNX export and compilation capabilities.
+
+    This class extends QEFFBaseModel to handle safety checker models with specific transformations and optimizations
+    for efficient inference on Qualcomm AI hardware. Safety checker models are commonly used in diffusion pipelines
+    to filter out potentially harmful or inappropriate generated content.
+    """
+
+    def __init__(self, model: nn.modules):
+        super().__init__(model.vae)
+        self.model = model.safety_checker
+
+    def export(
+        self,
+        inputs,
+        output_names,
+        dynamic_axes,
+        export_dir=None,
+        export_kwargs=None,
+    ):
+        return self._export(
+            example_inputs=inputs,
+            output_names=output_names,
+            dynamic_axes=dynamic_axes,
+            export_dir=export_dir,
+            export_kwargs=export_kwargs,
+        )
+
+    def compile(self, specializations, **compiler_options):
+        self._compile(specializations=specializations, **compiler_options)
+
+    @property
+    def model_hash(self) -> str:
+        # Compute the hash with: model_config, continuous_batching, transforms
+        mhash = hashlib.sha256()
+        mhash.update(to_hashable(self.model.config.to_diff_dict()))
+        mhash.update(to_hashable(self._transform_names()))
+        mhash = mhash.hexdigest()[:16]
+        return mhash
+
+    @property
+    def model_name(self) -> str:
+        mname = self.model.__class__.__name__
+        if mname.startswith("QEff") or mname.startswith("QEFF"):
+            mname = mname[4:]
+        return mname
 
 
 class QEffFluxTransformerModel(QEFFBaseModel):
-    """
-    Wrapper for Flux Transformer2D models with ONNX export and QAIC compilation capabilities.
-
-    This class handles Flux Transformer2D models with specific transformations and optimizations
-    for efficient inference on Qualcomm AI hardware. Flux uses a transformer-based diffusion
-    architecture instead of traditional UNet, with dual transformer blocks and adaptive layer
-    normalization (AdaLN) for conditioning.
-
-    Attributes:
-        model (nn.Module): The wrapped Flux transformer model
-        _pytorch_transforms (List): PyTorch transformations applied before ONNX export
-        _onnx_transforms (List): ONNX transformations applied after export
-    """
-
-    _pytorch_transforms = [AttentionTransform, NormalizationTransform, CustomOpsTransform]
+    _pytorch_transforms = [AttentionTransform, CustomOpsTransform, NormalizationTransform]
     _onnx_transforms = [FP16ClipTransform, SplitTensorsTransform]
+    """
+    QEffFluxTransformerModel is a wrapper class for Flux Transformer2D models that provides ONNX export and compilation capabilities.
 
-    @property
-    def get_model_config(self) -> Dict:
-        """
-        Get the model configuration as a dictionary.
+    This class extends QEFFBaseModel to handle Flux Transformer2D models with specific transformations and optimizations
+    for efficient inference on Qualcomm AI hardware. It is designed for the newer Flux transformer architecture
+    that uses transformer-based diffusion models instead of traditional UNet architectures.
+    """
 
-        Returns:
-            Dict: The configuration dictionary of the underlying Flux transformer model
-        """
-        return self.model.config.__dict__
-
-    def __init__(self, model: nn.Module) -> None:
-        """
-        Initialize the Flux transformer wrapper.
-
-        Args:
-            model (nn.Module): The Flux transformer model to wrap
-            use_onnx_subfunctions (bool): Whether to export transformer blocks as ONNX functions
-                                     for better modularity and potential optimization
-        """
+    def __init__(self, model: nn.modules, use_onnx_function):
         super().__init__(model)
+        if use_onnx_function:
+            self._pytorch_transforms.append(OnnxFunctionTransform)
+            model, _ = OnnxFunctionTransform.apply(model)
+        self.model = model
 
-    def get_onnx_params(
-        self,
-        batch_size: int = constants.ONNX_EXPORT_EXAMPLE_BATCH_SIZE,
-        seq_length: int = constants.FLUX_ONNX_EXPORT_SEQ_LENGTH,
-        cl: int = constants.FLUX_ONNX_EXPORT_COMPRESSED_LATENT_DIM,
-    ) -> Tuple[Dict, Dict, List[str]]:
-        """
-        Generate ONNX export configuration for the Flux transformer.
-
-        Creates example inputs for all Flux-specific inputs including hidden states,
-        text embeddings, timestep conditioning, and AdaLN embeddings.
-
-        Args:
-            batch_size (int): Batch size for example inputs (default: FLUX_ONNX_EXPORT_BATCH_SIZE)
-            seq_length (int): Text sequence length (default: FLUX_ONNX_EXPORT_SEQ_LENGTH)
-            cl (int): Compressed latent dimension (default: FLUX_ONNX_EXPORT_COMPRESSED_LATENT_DIM)
-
-        Returns:
-            Tuple containing:
-                - example_inputs (Dict): Sample inputs for ONNX export
-                - dynamic_axes (Dict): Specification of dynamic dimensions
-                - output_names (List[str]): Names of model outputs
-        """
+    def get_onnx_config(self, batch_size=1, seq_length=256, cl=4096):
         example_inputs = {
-            # Latent representation of the image
             "hidden_states": torch.randn(batch_size, cl, self.model.config.in_channels, dtype=torch.float32),
             "encoder_hidden_states": torch.randn(
                 batch_size, seq_length, self.model.config.joint_attention_dim, dtype=torch.float32
@@ -398,30 +294,15 @@ class QEffFluxTransformerModel(QEFFBaseModel):
             "timestep": torch.tensor([1.0], dtype=torch.float32),
             "img_ids": torch.randn(cl, 3, dtype=torch.float32),
             "txt_ids": torch.randn(seq_length, 3, dtype=torch.float32),
-            # AdaLN embeddings for dual transformer blocks
-            # Shape: [num_layers, FLUX_ADALN_DUAL_BLOCK_CHUNKS, FLUX_ADALN_HIDDEN_DIM]
             "adaln_emb": torch.randn(
-                self.model.config["num_layers"],
-                constants.FLUX_ADALN_DUAL_BLOCK_CHUNKS,
-                constants.FLUX_ADALN_HIDDEN_DIM,
-                dtype=torch.float32,
-            ),
-            # AdaLN embeddings for single transformer blocks
-            # Shape: [num_single_layers, FLUX_ADALN_SINGLE_BLOCK_CHUNKS, FLUX_ADALN_HIDDEN_DIM]
-            "adaln_single_emb": torch.randn(
-                self.model.config["num_single_layers"],
-                constants.FLUX_ADALN_SINGLE_BLOCK_CHUNKS,
-                constants.FLUX_ADALN_HIDDEN_DIM,
-                dtype=torch.float32,
-            ),
-            # Output AdaLN embedding
-            # Shape: [batch_size, FLUX_ADALN_OUTPUT_DIM] for final projection
-            "adaln_out": torch.randn(batch_size, constants.FLUX_ADALN_OUTPUT_DIM, dtype=torch.float32),
+                self.model.config.num_layers, 12, 3072, dtype=torch.float32
+            ),  # num_layers, #chunks, # Adalan_hidden_dim
+            "adaln_single_emb": torch.randn(self.model.config.num_single_layers, 3, 3072, dtype=torch.float32),
+            "adaln_out": torch.randn(batch_size, 6144, dtype=torch.float32),
         }
 
         output_names = ["output"]
 
-        # Define dynamic dimensions for runtime flexibility
         dynamic_axes = {
             "hidden_states": {0: "batch_size", 1: "cl"},
             "encoder_hidden_states": {0: "batch_size", 1: "seq_len"},
@@ -434,48 +315,52 @@ class QEffFluxTransformerModel(QEFFBaseModel):
 
     def export(
         self,
-        inputs: Dict,
-        output_names: List[str],
-        dynamic_axes: Dict,
-        export_dir: str = None,
-        export_kwargs: Dict = None,
-        use_onnx_subfunctions: bool = False,
-    ) -> str:
-        """
-        Export the Flux transformer model to ONNX format.
-
-        Args:
-            inputs (Dict): Example inputs for ONNX export
-            output_names (List[str]): Names of model outputs
-            dynamic_axes (Dict): Specification of dynamic dimensions
-            export_dir (str, optional): Directory to save ONNX model
-            export_kwargs (Dict, optional): Additional export arguments (e.g., export_modules_as_functions)
-
-        Returns:
-            str: Path to the exported ONNX model
-        """
-
-        if use_onnx_subfunctions:
-            export_kwargs = {"export_modules_as_functions": {QEffFluxTransformerBlock, QEffFluxSingleTransformerBlock}}
-
-        # Sort _use_default_values in config to ensure consistent hash generation during export
-        self.model.config["_use_default_values"].sort()
-
+        inputs,
+        output_names,
+        dynamic_axes,
+        export_dir=None,
+        export_kwargs=None,
+    ):
         return self._export(
             example_inputs=inputs,
             output_names=output_names,
             dynamic_axes=dynamic_axes,
             export_dir=export_dir,
             export_kwargs=export_kwargs,
-            offload_pt_weights=False,  # As weights are needed with AdaLN changes
         )
 
-    def compile(self, specializations: List[Dict], **compiler_options) -> None:
-        """
-        Compile the ONNX model for Qualcomm AI hardware.
+    def get_specializations(self, batch_size: int, seq_len: int, cl: int):
+        specializations = [
+            {
+                "batch_size": batch_size,
+                "stats-batchsize": batch_size,
+                "num_layers": self.model.config.num_layers,
+                "num_single_layers": self.model.config.num_single_layers,
+                "seq_len": seq_len,
+                "cl": cl,
+                "steps": 1,
+            }
+        ]
 
-        Args:
-            specializations (List[Dict]): Model specialization configurations
-            **compiler_options: Additional compiler options (e.g., num_cores, aic_num_of_activations)
-        """
+        return specializations
+
+    def compile(self, specializations, **compiler_options):
         self._compile(specializations=specializations, **compiler_options)
+
+    @property
+    def model_hash(self) -> str:
+        # Compute the hash with: model_config, continuous_batching, transforms
+        mhash = hashlib.sha256()
+        dict_model_config = dict(self.model.config)
+        dict_model_config.pop("_use_default_values", None)
+        mhash.update(to_hashable(dict_model_config))
+        mhash.update(to_hashable(self._transform_names()))
+        mhash = mhash.hexdigest()[:16]
+        return mhash
+
+    @property
+    def model_name(self) -> str:
+        mname = self.model.__class__.__name__
+        if mname.startswith("QEff") or mname.startswith("QEFF"):
+            mname = mname[4:]
+        return mname
