+++ QEfficient/diffusers/pipelines/pipeline_utils.py
@@ -5,362 +5,96 @@
 #
 # ----------------------------------------------------------------------------
 
-import copy
-import hashlib
+import os
+from dataclasses import dataclass
+from typing import TYPE_CHECKING, List, Optional, Union
 
-import torch
-import torch.nn as nn
+import numpy as np
+import PIL.Image
 
-from QEfficient.base.modeling_qeff import QEFFBaseModel
-from QEfficient.base.onnx_transforms import FP16ClipTransform, SplitTensorsTransform
-from QEfficient.diffusers.models.pytorch_transforms import (
-    AttentionTransform,
-    CustomOpsTransform,
-    NormalizationTransform,
-    OnnxFunctionTransform,
-)
-from QEfficient.transformers.models.pytorch_transforms import (
-    T5ModelTransform,
-)
-from QEfficient.utils import constants
-from QEfficient.utils.cache import to_hashable
+from QEfficient.utils._utils import load_json
+
+if TYPE_CHECKING:
+    from QEfficient.diffusers.pipelines.flux.pipeline_flux import QEFFFluxPipeline
 
 
-class QEffTextEncoder(QEFFBaseModel):
-    _pytorch_transforms = [CustomOpsTransform, T5ModelTransform]
-    _onnx_transforms = [FP16ClipTransform, SplitTensorsTransform]
+def config_manager(cls, config_source: Optional[str] = None):
     """
-    QEffTextEncoder is a wrapper class for text encoder models that provides ONNX export and compilation capabilities.
+    JSON-based compilation configuration manager for diffusion pipelines.
 
-    This class extends QEFFBaseModel to handle text encoder models (like T5EncoderModel) with specific
-    transformations and optimizations for efficient inference on Qualcomm AI hardware.
+    Supports loading configuration from JSON files only. Automatically detects
+    model type and handles model-specific requirements.
+    Initialize the configuration manager.
+
+    Args:
+        config_source: Path to JSON configuration file. If None, uses default config.
     """
+    if config_source is None:
+        config_source = cls.get_default_config_path()
 
-    def __init__(self, model: nn.modules):
-        super().__init__(model)
-        self.model = copy.deepcopy(model)
+    if not isinstance(config_source, str):
+        raise ValueError("config_source must be a path to JSON configuration file")
 
-    def get_onnx_config(self):
-        bs = constants.ONNX_EXPORT_EXAMPLE_BATCH_SIZE
+    # Direct use of load_json utility - no wrapper needed
+    if not os.path.exists(config_source):
+        raise FileNotFoundError(f"Configuration file not found: {config_source}")
 
-        example_inputs = {
-            "input_ids": torch.zeros((bs, self.model.config.max_position_embeddings), dtype=torch.int64),
-        }
-
-        dynamic_axes = {"input_ids": {0: "batch_size", 1: "seq_len"}}
-        output_names = ["last_hidden_state", "pooler_output"]
-
-        if self.model.__class__.__name__ == "T5EncoderModel":
-            output_names = ["last_hidden_state"]
-        else:
-            example_inputs["output_hidden_states"] = False
-
-        return example_inputs, dynamic_axes, output_names
-
-    def export(
-        self,
-        inputs,
-        output_names,
-        dynamic_axes,
-        export_dir=None,
-        export_kwargs=None,
-    ):
-        return self._export(
-            example_inputs=inputs,
-            output_names=output_names,
-            dynamic_axes=dynamic_axes,
-            export_dir=export_dir,
-            export_kwargs=export_kwargs,
-        )
-
-    def compile(self, specializations, **compiler_options):
-        self._compile(specializations=specializations, **compiler_options)
-
-    @property
-    def model_hash(self) -> str:
-        # Compute the hash with: model_config, continuous_batching, transforms
-        mhash = hashlib.sha256()
-        mhash.update(to_hashable(self.model.config.to_diff_dict()))
-        mhash.update(to_hashable(self._transform_names()))
-        mhash = mhash.hexdigest()[:16]
-        return mhash
-
-    @property
-    def model_name(self) -> str:
-        mname = self.model.__class__.__name__
-        if mname.startswith("QEff") or mname.startswith("QEFF"):
-            mname = mname[4:]
-        return mname
+    cls.custom_config = load_json(config_source)
 
 
-class QEffUNet(QEFFBaseModel):
-    _pytorch_transforms = [CustomOpsTransform]
-    _onnx_transforms = [FP16ClipTransform, SplitTensorsTransform]
-
+def set_module_device_ids(cls):
     """
-    QEffUNet is a wrapper class for UNet models that provides ONNX export and compilation capabilities.
+    Set device IDs for each module based on the custom configuration.
 
-    This class extends QEFFBaseModel to handle UNet models with specific transformations and optimizations
-    for efficient inference on Qualcomm AI hardware. It is commonly used in diffusion models for image
-    generation tasks.
+    Iterates through all modules in the pipeline and assigns device IDs
+    from the configuration file to each module's device_ids attribute.
     """
-
-    def __init__(self, model: nn.modules):
-        super().__init__(model.unet)
-        self.model = model.unet
-
-    def export(
-        self,
-        inputs,
-        output_names,
-        dynamic_axes,
-        export_dir=None,
-        export_kwargs=None,
-    ):
-        return self._export(
-            example_inputs=inputs,
-            output_names=output_names,
-            dynamic_axes=dynamic_axes,
-            export_dir=export_dir,
-            export_kwargs=export_kwargs,
-        )
-
-    def compile(self, specializations, **compiler_options):
-        self._compile(specializations=specializations, **compiler_options)
-
-    @property
-    def model_hash(self) -> str:
-        # Compute the hash with: model_config, continuous_batching, transforms
-        mhash = hashlib.sha256()
-        mhash.update(to_hashable(dict(self.model.config)))
-        mhash.update(to_hashable(self._transform_names()))
-        mhash = mhash.hexdigest()[:16]
-        return mhash
-
-    @property
-    def model_name(self) -> str:
-        mname = self.model.__class__.__name__
-        if mname.startswith("QEff") or mname.startswith("QEFF"):
-            mname = mname[4:]
-        return mname
+    config_modules = cls.custom_config["modules"]
+    for module_name, module_obj in cls.has_module:
+        module_obj.device_ids = config_modules[module_name]["execute"]["device_ids"]
 
 
-class QEffVAE(QEFFBaseModel):
-    _pytorch_transforms = [CustomOpsTransform]
-    _onnx_transforms = [FP16ClipTransform, SplitTensorsTransform]
+@dataclass
+class QEffPipelineOutput:
+    pipeline: "QEFFFluxPipeline"
+    images: Union[List[PIL.Image.Image], np.ndarray]
+    E2E_time: int
 
-    """
-    QEffVAE is a wrapper class for Variational Autoencoder (VAE) models that provides ONNX export and compilation capabilities.
+    def __repr__(self):
+        output_str = "=" * 60 + "\n"
+        output_str += "QEfficient Diffusers Pipeline Inference Report\n"
+        output_str += "=" * 60 + "\n\n"
 
-    This class extends QEFFBaseModel to handle VAE models with specific transformations and optimizations
-    for efficient inference on Qualcomm AI hardware. VAE models are commonly used in diffusion pipelines
-    for encoding images to latent space and decoding latent representations back to images.
-    """
+        # End-to-End time
+        output_str += f"End-to-End Inference Time: {self.E2E_time:.4f} s\n\n"
 
-    def __init__(self, model: nn.modules, type: str):
-        super().__init__(model.vae)
-        self.model = copy.deepcopy(model.vae)
-        self.type = type
+        # Module-wise inference times
+        output_str += "Module-wise Inference Times:\n"
+        output_str += "-" * 60 + "\n"
 
-    def get_onnx_config(self, latent_height=32, latent_width=32):
-        # VAE decode
-        bs = constants.ONNX_EXPORT_EXAMPLE_BATCH_SIZE
-        example_inputs = {
-            "latent_sample": torch.randn(bs, 16, latent_height, latent_width),
-            "return_dict": False,
-        }
+        # Iterate through all modules using has_module
+        for module_name, module_obj in self.pipeline.has_module:
+            if hasattr(module_obj, "inference_time"):
+                inference_time = module_obj.inference_time
 
-        output_names = ["sample"]
+                # Format module name for display
+                display_name = module_name.replace("_", " ").title()
 
-        dynamic_axes = {
-            "latent_sample": {0: "batch_size", 1: "channels", 2: "latent_height", 3: "latent_width"},
-        }
-        return example_inputs, dynamic_axes, output_names
+                # Handle transformer specially as it has a list of times
+                if isinstance(inference_time, list) and len(inference_time) > 0:
+                    total_time = sum(inference_time)
+                    avg_time = total_time / len(inference_time)
+                    output_str += f"  {display_name:25s} {total_time:.4f} s\n"
+                    output_str += f"    - Total steps: {len(inference_time)}\n"
+                    output_str += f"    - Average per step:    {avg_time:.4f} s\n"
+                    output_str += f"    - Min step time:       {min(inference_time):.4f} s\n"
+                    output_str += f"    - Max step time:       {max(inference_time):.4f} s\n"
+                else:
+                    # Single inference time value
+                    output_str += f"  {display_name:25s} {inference_time:.4f} s\n"
 
-    def export(
-        self,
-        inputs,
-        output_names,
-        dynamic_axes,
-        export_dir=None,
-        export_kwargs=None,
-    ):
-        return self._export(
-            example_inputs=inputs,
-            output_names=output_names,
-            dynamic_axes=dynamic_axes,
-            export_dir=export_dir,
-            export_kwargs=export_kwargs,
-        )
+        output_str += "-" * 60 + "\n\n"
 
-    def compile(self, specializations, **compiler_options):
-        self._compile(specializations=specializations, **compiler_options)
+        output_str += "=" * 60 + "\n"
 
-    @property
-    def model_hash(self) -> str:
-        # Compute the hash with: model_config, continuous_batching, transforms
-        mhash = hashlib.sha256()
-        mhash.update(to_hashable(dict(self.model.config)))
-        mhash.update(to_hashable(self._transform_names()))
-        mhash.update(to_hashable(self.type))
-        mhash = mhash.hexdigest()[:16]
-        return mhash
-
-    @property
-    def model_name(self) -> str:
-        mname = self.model.__class__.__name__
-        if mname.startswith("QEff") or mname.startswith("QEFF"):
-            mname = mname[4:]
-        return mname
-
-
-class QEffSafetyChecker(QEFFBaseModel):
-    _pytorch_transforms = [CustomOpsTransform]
-    _onnx_transforms = [FP16ClipTransform, SplitTensorsTransform]
-
-    """
-    QEffSafetyChecker is a wrapper class for safety checker models that provides ONNX export and compilation capabilities.
-
-    This class extends QEFFBaseModel to handle safety checker models with specific transformations and optimizations
-    for efficient inference on Qualcomm AI hardware. Safety checker models are commonly used in diffusion pipelines
-    to filter out potentially harmful or inappropriate generated content.
-    """
-
-    def __init__(self, model: nn.modules):
-        super().__init__(model.vae)
-        self.model = model.safety_checker
-
-    def export(
-        self,
-        inputs,
-        output_names,
-        dynamic_axes,
-        export_dir=None,
-        export_kwargs=None,
-    ):
-        return self._export(
-            example_inputs=inputs,
-            output_names=output_names,
-            dynamic_axes=dynamic_axes,
-            export_dir=export_dir,
-            export_kwargs=export_kwargs,
-        )
-
-    def compile(self, specializations, **compiler_options):
-        self._compile(specializations=specializations, **compiler_options)
-
-    @property
-    def model_hash(self) -> str:
-        # Compute the hash with: model_config, continuous_batching, transforms
-        mhash = hashlib.sha256()
-        mhash.update(to_hashable(self.model.config.to_diff_dict()))
-        mhash.update(to_hashable(self._transform_names()))
-        mhash = mhash.hexdigest()[:16]
-        return mhash
-
-    @property
-    def model_name(self) -> str:
-        mname = self.model.__class__.__name__
-        if mname.startswith("QEff") or mname.startswith("QEFF"):
-            mname = mname[4:]
-        return mname
-
-
-class QEffFluxTransformerModel(QEFFBaseModel):
-    _pytorch_transforms = [AttentionTransform, CustomOpsTransform, NormalizationTransform]
-    _onnx_transforms = [FP16ClipTransform, SplitTensorsTransform]
-    """
-    QEffFluxTransformerModel is a wrapper class for Flux Transformer2D models that provides ONNX export and compilation capabilities.
-
-    This class extends QEFFBaseModel to handle Flux Transformer2D models with specific transformations and optimizations
-    for efficient inference on Qualcomm AI hardware. It is designed for the newer Flux transformer architecture
-    that uses transformer-based diffusion models instead of traditional UNet architectures.
-    """
-
-    def __init__(self, model: nn.modules, use_onnx_function):
-        super().__init__(model)
-        if use_onnx_function:
-            self._pytorch_transforms.append(OnnxFunctionTransform)
-            model, _ = OnnxFunctionTransform.apply(model)
-        self.model = model
-
-    def get_onnx_config(self, batch_size=1, seq_length=256, cl=4096):
-        example_inputs = {
-            "hidden_states": torch.randn(batch_size, cl, self.model.config.in_channels, dtype=torch.float32),
-            "encoder_hidden_states": torch.randn(
-                batch_size, seq_length, self.model.config.joint_attention_dim, dtype=torch.float32
-            ),
-            "pooled_projections": torch.randn(batch_size, self.model.config.pooled_projection_dim, dtype=torch.float32),
-            "timestep": torch.tensor([1.0], dtype=torch.float32),
-            "img_ids": torch.randn(cl, 3, dtype=torch.float32),
-            "txt_ids": torch.randn(seq_length, 3, dtype=torch.float32),
-            "adaln_emb": torch.randn(
-                self.model.config.num_layers, 12, 3072, dtype=torch.float32
-            ),  # num_layers, #chunks, # Adalan_hidden_dim
-            "adaln_single_emb": torch.randn(self.model.config.num_single_layers, 3, 3072, dtype=torch.float32),
-            "adaln_out": torch.randn(batch_size, 6144, dtype=torch.float32),
-        }
-
-        output_names = ["output"]
-
-        dynamic_axes = {
-            "hidden_states": {0: "batch_size", 1: "cl"},
-            "encoder_hidden_states": {0: "batch_size", 1: "seq_len"},
-            "pooled_projections": {0: "batch_size"},
-            "timestep": {0: "steps"},
-            "img_ids": {0: "cl"},
-        }
-
-        return example_inputs, dynamic_axes, output_names
-
-    def export(
-        self,
-        inputs,
-        output_names,
-        dynamic_axes,
-        export_dir=None,
-        export_kwargs=None,
-    ):
-        return self._export(
-            example_inputs=inputs,
-            output_names=output_names,
-            dynamic_axes=dynamic_axes,
-            export_dir=export_dir,
-            export_kwargs=export_kwargs,
-        )
-
-    def get_specializations(self, batch_size: int, seq_len: int, cl: int):
-        specializations = [
-            {
-                "batch_size": batch_size,
-                "stats-batchsize": batch_size,
-                "num_layers": self.model.config.num_layers,
-                "num_single_layers": self.model.config.num_single_layers,
-                "seq_len": seq_len,
-                "cl": cl,
-                "steps": 1,
-            }
-        ]
-
-        return specializations
-
-    def compile(self, specializations, **compiler_options):
-        self._compile(specializations=specializations, **compiler_options)
-
-    @property
-    def model_hash(self) -> str:
-        # Compute the hash with: model_config, continuous_batching, transforms
-        mhash = hashlib.sha256()
-        dict_model_config = dict(self.model.config)
-        dict_model_config.pop("_use_default_values", None)
-        mhash.update(to_hashable(dict_model_config))
-        mhash.update(to_hashable(self._transform_names()))
-        mhash = mhash.hexdigest()[:16]
-        return mhash
-
-    @property
-    def model_name(self) -> str:
-        mname = self.model.__class__.__name__
-        if mname.startswith("QEff") or mname.startswith("QEFF"):
-            mname = mname[4:]
-        return mname
+        return output_str
