+++ examples/diffusers/flux/flux_1_schnell.py
@@ -8,14 +8,16 @@ import torch
 
 from QEfficient import QEFFFluxPipeline
 
-pipeline = QEFFFluxPipeline.from_pretrained("black-forest-labs/FLUX.1-schnell")
+pipeline = QEFFFluxPipeline.from_pretrained("black-forest-labs/FLUX.1-schnell", use_onnx_function=True)
 
-image = pipeline(
+output = pipeline(
     prompt="A cat holding a sign that says hello world",
     guidance_scale=0.0,
     num_inference_steps=4,
     max_sequence_length=256,
     generator=torch.manual_seed(42),
-).images[0]
+)
+image = pipeline.images[0]
+image.save("cat_with_sign.png")
 
-image.save("flux-schnell_aic_1024.png")
+print(output)
